tmdb_wall/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
httpx>=0.27.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 夸克搜索服务依赖
sqlalchemy>=2.0.23
aiohttp>=3.9.1
requests>=2.31.0
